display: move `Font` to `components/font`
