+++ esphome/components/font/font.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include "esphome/core/datatypes.h"
-#include "display_buffer.h"
+#include "esphome/core/color.h"
+#include "esphome/components/display/display_buffer.h"
 
 namespace esphome {
-namespace display {
+namespace font {
 
-class DisplayBuffer;
 class Font;
 
 struct GlyphData {
@@ -22,7 +22,7 @@ class Glyph {
  public:
   Glyph(const GlyphData *data) : glyph_data_(data) {}
 
-  void draw(int x, int y, DisplayBuffer *display, Color color) const;
+  void draw(int x, int y, display::DisplayBuffer *display, Color color) const;
 
   const char *get_char() const;
 
@@ -38,7 +38,7 @@ class Glyph {
   const GlyphData *glyph_data_;
 };
 
-class Font : public BaseFont {
+class Font : public display::BaseFont {
  public:
   /** Construct the font with the given glyphs.
    *
@@ -50,7 +50,7 @@ class Font : public BaseFont {
 
   int match_next_glyph(const char *str, int *match_length);
 
-  void print(int x_start, int y_start, DisplayBuffer *display, Color color, const char *text) override;
+  void print(int x_start, int y_start, display::DisplayBuffer *display, Color color, const char *text) override;
   void measure(const char *str, int *width, int *x_offset, int *baseline, int *height) override;
   inline int get_baseline() { return this->baseline_; }
   inline int get_height() { return this->height_; }
@@ -63,5 +63,5 @@ class Font : public BaseFont {
   int height_;
 };
 
-}  // namespace display
+}  // namespace font
 }  // namespace esphome
